refactor(app): create and manage QTs through the /protocols view
