+++ app/src/pages/ODD/ProtocolSetup/index.tsx
@@ -738,9 +738,6 @@ function PrepareToRun({
       {showConfirmCancelModal ? (
         <ConfirmCancelRunModal
           runId={runId}
-          isQuickTransfer={
-            protocolRecord?.data.protocolKind === 'quick-transfer'
-          }
           setShowConfirmCancelRunModal={setShowConfirmCancelModal}
           isActiveRun={false}
           protocolId={protocolId}
